Mobilize + componentize Footer
