--- src/components/Footer.js
@@ -1,245 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-
-import {
-  Button,
-  FormControl,
-  Grid,
-  Input,
-  Typography
-} from '@material-ui/core';
-import { withStyles } from '@material-ui/core/styles';
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-import '../assets/css/App.css';
-import codeforafrica from '../assets/images/logos/cfafrica_white.png';
-import innovateafrica from '../assets/images/logos/innovateafrica_white.png';
-
-const styles = theme => ({
-  root: {
-    flexGrow: 1,
-    color: 'white',
-    backgroundColor: '#424143'
-  },
-  titles: {
-    color: 'white'
-  },
-  fa: {
-    padding: '0.5rem',
-    transition: 'all .5s ease-in-out',
-    '&:hover': {
-      transform: 'scale(1.3)',
-      color: '#f3f33'
-    }
-  },
-  connectContainer: {
-    paddingTop: theme.spacing.unit * 5,
-    paddingBottom: theme.spacing.unit * 5,
-    paddingLeft: theme.spacing.unit * 5,
-    paddingRight: theme.spacing.unit * 5
-  },
-  icons: {
-    paddingTop: theme.spacing.unit * 1,
-    paddingBottom: theme.spacing.unit * 2
-  },
-  aboutContainer: {
-    textAlign: 'center',
-    padding: theme.spacing.unit * 3,
-    borderLeft: '1px solid white',
-    borderRight: '1px solid white'
-  },
-  footerButton: {
-    color: 'white',
-    backgroundColor: '#2A2A2B',
-    borderRadius: '0',
-    '&:hover': {
-      color: '#424143'
-    }
-  },
-  footerInput: {
-    width: '190px',
-    backgroundColor: 'white',
-    border: '1px solid white'
-  },
-  logoText: {
-    textAlign: 'center',
-    paddingTop: theme.spacing.unit * 4,
-    paddingLeft: theme.spacing.unit * 3
-  },
-  logoItems: {
-    textAlign: 'center',
-    paddingLeft: theme.spacing.unit * 3
-  },
-  links: {
-    color: 'white'
-  },
-  buttonLink: {
-    textDecoration: 'none',
-    color: 'white',
-    '&:hover': {
-      color: '#424143'
-    }
-  }
-});
-
-class Footer extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = { value: 'you@gmail.com' };
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange(e) {
-    this.setState({ value: e.target.value });
-  }
-
-  render() {
-    const { classes } = this.props;
-    const { value } = this.state;
-
-    return (
-      <Grid container item xs={12} className={classes.root}>
-        <Grid
-          container
-          spacing={0}
-          direction="row"
-          justify="center"
-          alignitems="center"
-          className={classes.connectContainer}
-        >
-          <Grid item xs={4} align="center">
-            <Typography variant="title" className={classes.titles}>
-              CONNECT WITH US
-            </Typography>
-            <div
-              container
-              direction="row"
-              justify="center"
-              alignitems="center"
-              className={classes.icons}
-            >
-              <a
-                href="https://www.facebook.com/sensorsAFRICA/#sensorsAFRICA"
-                className={classes.links}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <FontAwesomeIcon
-                  className={classes.fa}
-                  icon={['fab', 'facebook-f']}
-                  size="lg"
-                />
-              </a>
-              <a
-                href="https://twitter.com/sensorsAFRICA/#sensorsAFRICA"
-                className={classes.links}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <FontAwesomeIcon
-                  className={classes.fa}
-                  icon={['fab', 'twitter']}
-                  size="lg"
-                />
-              </a>
-              {/* <a href="" className={classes.links}>
-                <FontAwesomeIcon
-                  className={classes.fa}
-                  icon={['fab', 'instagram']}
-                  size="lg"
-                />
-              </a> */}
-            </div>
-            <Grid
-              container
-              direction="column"
-              justify="space-around"
-              alignitems="center"
-              className={classes.footerMargin}
-            >
-              <FormControl style={{ margin: '0 auto' }}>
-                <Input
-                  type="text"
-                  id="you@gmail.com"
-                  value={value}
-                  onChange={this.handleChange}
-                  className={classes.footerInput}
-                />
-                <p>
-                  <a
-                    href="https://codeforafrica.us6.list-manage.com/subscribe/post?u=65e5825507b3cec760f272e79&id=c2ff751541"
-                    className={classes.buttonLink}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    <Button
-                      variant="contained"
-                      className={classes.footerButton}
-                    >
-                      SUBSCRIBE TO UPDATES
-                    </Button>
-                  </a>
-                </p>
-              </FormControl>
-            </Grid>
-          </Grid>
-          <Grid item xs={4} className={classes.aboutContainer}>
-            <Typography variant="title" className={classes.titles}>
-              ABOUT SENSORS.AFRICA
-            </Typography>
-            <p>
-              sensors.AFRICA is a pan-African citizen science initiative that
-              uses sensors to monitor air, water and sound polution to give
-              citizens actionable information about their cities.
-            </p>
-            <a
-              href="https://medium.com/code-for-africa/tagged/innovateafrica"
-              className={classes.buttonLink}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Button variant="contained" className={classes.footerButton}>
-                READ MORE
-              </Button>
-            </a>
-          </Grid>
-          <Grid item xs={4}>
-            <p className={classes.logoText}>
-              This initiative was seed-funded by innovateAFRICA, and is being
-              incubated by Code for Africa
-            </p>
-            <Grid
-              container
-              direction="row"
-              justify="center"
-              alignItems="center"
-            >
-              <a
-                href="https://codeforafrica.org/"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img src={codeforafrica} alt="Code For Africa" />
-              </a>
-              <a
-                href="https://www.innoafrica.org/"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img src={innovateafrica} alt="Innovate Africa" />
-              </a>
-            </Grid>
-          </Grid>
-        </Grid>
-      </Grid>
-    );
-  }
-}
-
-Footer.propTypes = {
-  classes: PropTypes.object.isRequired
-};
-
-export default withStyles(styles)(Footer);
